Fix package manager call for RHEL-based distros
Assuming users will be working with latests versions
for RHEL, Centos or Fedora, the requirement for EPEL
repo is not longer needed. This requirement was due
to argparse not being present in standard repos, something
that does not hold anymore.
For this reason, this patch set removes the installation
of the EPEL repo.
On the other hand, this patch set also makes an extra check
to use dnf when this script is run on Fedora. Since Fedora 22
dnf has been the default package manager instead of yum.

Change-Id: I9fcd47a0bf9302506949fa56efb42b77552710f6
Closes-Bug: #1611527