Removes LXC/LXD support on manila-image-elements
The support for LXC/LXD was added as part of the initial effort
to develop a container driver for Manila which was later on
abandoned. Despite of that, the code in the manila-image-elements
remained.

This patch set removes LXC/LXD container support in
manila-image-elements. Code-path used for generation of Docker images is
preserved.

This will help for the parallel effort of making
manila-image-elements usable by end users.

Change-Id: I99c6a7fbf2b2fb0cebb7b16d94c66f807e7ac7ef
Closes-Bug: #1646615