Suppress some bashate warnings
When running ``tox`` for manila image elements bashate warnings
are emitted for lines that are too long.  Suppress these since
we have no intention of enforcing that requirement.  Also add a
shebang to the top of a shell script, which does no harm (even
though it is not needed when sourcing the script to set env files)
in order to make another warning go away.

TrivialFix

Change-Id: I7f2f2723f99d64a0e28b4d0fe43b79e7201f6121