doc migration: new directory layout
This patch introduces a new directory layout
in doc/source in conformance with the OpenStack
manuals project migration spec [1], moves the
existing content in manila/doc/source into the
new directories, and adjusts index files accordingly.

This is the first step in the migration process
as outlined in the spec.

[1] https://specs.openstack.org/openstack/docs-specs/specs/pike/os-manuals-migration.html

Partial-Bug: #1706181
Change-Id: I6b876f6e697659a5fe7046d5c5d4fd4ed98f9828
Needed-By: If13ebd1ede6fbc997d74425ec78e15621f1f80b2
Depends-On: I834fbef9783739fd664246ff710df1ab42ce2af5
Depends-On: Ia750cb049c0f53a234ea70ce1f2bbbb7a2aa9454