Add native Zuul v3 manila image elements jobs
Replace the auto-converted legacy jobs with native jobs
and define all the jobs in-tree.  After we merge this we
can remove the legacy-manila-publish-generic job from the
openstack-zuul-jobs repository.

Note that the old job apparently relied on some JJB magic
to construct tagged-image names.  This no longer works so
instead we use 'git describe --tags' to obtain a current
tag, publish the image with its tagged name, and symlink
manila-service-image-master.qcow2 -- which is the expected
image-name for our CI jobs -- to it.

Unfortunately the new publish job can't be tested
beforehand so we may have to follow this with another
patch if there are issues.

Closes-Bug: #1812478
Change-Id: Ie0c367591c0e1ae4fefb62688ecd76aff5b94464