Modify job to build images
We currently only build the generic driver
image with the job "manila-image-elements-buildimage-generic".
Since we'd like to use a cephfs based client image,
we can reuse the job to produce two images instead
of one. The new job "manila-image-elements-buildimages"
produces two images: "manila-service-image-cephfs-{tag}.qcow2"
and "manila-service-image-{tag}.qcow2"; where the
tag corresponds to the version of this project.

The job also creates a soft link of the tagged images
with a "-master" suffix instead of the tag - this
is useful for CI jobs that don't need to worry
about changing/using the latest tag.

Change-Id: I60c6b38b5d5de7649c0d760519a79288fafca1e3
Signed-off-by: Goutham Pacha Ravi <gouthampravi@gmail.com>