Refactor cephfs element
DIB ignores a plain "install.d" directive
silently currently, producing images without
the packages we asked for.

Use the "package-installs" approach to install
element dependencies. No configuration is added
at the moment, but may be required in the future.

Also install the "fuse" package to get the ability
to execute "fusermount" command.

Change-Id: Ifbd7344b20940575348a856521539425b34a4279
Signed-off-by: Goutham Pacha Ravi <gouthampravi@gmail.com>