Native Zuulv3 container scenario test job
Replace the "manila-tempest-dsvm-container-scenario-custom-image"
job with a native Zuulv3 style job that does not use
the deprecated devstack-gate project. This new test
job runs backend tests at the moment with the container
driver. Scenario tests that include mounting shares to
tenant VMs aren't yet enabled, and doing so is a TODO
item for the future.

Change-Id: I65379ffcf06e9de111d69cd3b5bc4acc7af0547e
Signed-off-by: Goutham Pacha Ravi <gouthampravi@gmail.com>